requirements.txt
transformers==4.39.2
flask==3.0.2
requests==2.31.0
Flask-Cors==4.0.0
sentencepiece==0.2.0
pymongo==4.7.2
python-dotenv==1.0.1
flask-pymongo==2.3.0